requirements.txt
Flask==2.3.3
pandas==2.0.3
qrcode[pil]==7.4.2
Pillow==10.0.0